@@ -21,11 +21,14 @@
     "webpack": "^1.15.0"
   },
   "dependencies": {
+    "babel": "^6.23.0",
+    "babel-core": "^6.24.1",
+    "create-react-class": "^15.5.2",
     "fs": "0.0.2",
     "pixi.js": "^3.0.11",
-    "react": "^15.5.4",
+    "react": "^16.0.0-alpha.6",
     "react-color": "^2.2.0",
-    "react-dom": "^15.1.0",
+    "react-dom": "^16.0.0-alpha.6",
     "react-emoji": "^0.4.4",
     "react-emoji-picker": "^1.0.12",
     "react-native": "^0.43.4",
