@@ -4,25 +4,28 @@ import ReactEmoji 		from 'react-emoji';
 import EmojiPicker 		from 'react-emoji-picker';
 import emojiMap 		from '../../node_modules/react-emoji-picker/lib/emojiMap';
 
-export default React.createClass({
+export default class extends React.Component {
+	constructor(props) {
+		super(props);
+	}
 	mixins: [
 	    ReactEmoji
-	],
+	]
 	handleColorPick(color) {
 		this.props.player.color 					= color.hex.replace('#', '0x')
 		this.props.game.user_token.style.background = color.hex
 		this.props.player.color_obj 				= color;
-	},
+	}
 	closeModal()
 	{
 		this.props.game.modal_container.style.display = 'none';
-	},
+	}
 	setEmoji(emoji)
 	{
 		this.props.player.emoji = emoji;
 		this.props.player.emoji_img = this.emojify(emoji)[0].props.src
 		this.props.game.user_token.innerHTML = '<img src="'+this.props.player.emoji_img+'" style="width:80px">';
-	},
+	}
 	render() {
 		var self = this
 		var saveCustomizing = function()
@@ -84,11 +87,11 @@ export default React.createClass({
 		        type="button" 
 		        style={{margin:"10px 0"}} 
 		        className="btn btn-primary">
-		        Save changes
+		        	Save changes
 		        </button>
 		      </div>
 		    </div>
 		  </div>
 		</div>;
 	}
-});
+}
